move loadOlm to init.ts
